backend/requirements.txt
fastapi
uvicorn
numpy
torch
pydantic
pydantic-settings
sqlalchemy
asyncpg
python-dotenv
google-generativeai